legacy_v07/backend-services/services/brain-orchestrator/src/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
PyJWT==2.11.0
requests==2.31.0
web3==7.14.0
google-cloud-pubsub==2.21.3
google-cloud-storage==2.16.0
google-cloud-bigquery==3.23.1
google-cloud-bigtable==2.24.0
google-cloud-secret-manager==2.20.0
circuitbreaker==2.0.0
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0
